requirements.txt
nebula3-python
mysql
mysql-connector-python
sqlparse
